as per feedback in today SIG
simplified the implementation to remove unnecessary interfaces.
