+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/export/__init__.py
@@ -178,49 +178,3 @@ class PeriodicExportingMetricReader(MetricReader):
         self._daemon_thread.join()
         self._exporter.shutdown()
         return True
-
-
-class PullMetricExporter(MetricExporter):
-    def __init__(self) -> None:
-        self._collect: Callable[
-            ["PullMetricExporter"], None,
-        ] = None
-
-    @final
-    def _set_collect_callback(
-        self,
-        func: Callable[
-            ["PullMetricExporter"], None,
-        ],
-    ) -> None:
-        """This function is internal to the SDK. It should not be called or overriden by users"""
-        self._collect = func
-
-    @final
-    def collect(self) -> None:
-        if self._collect is not None:
-            self._collect()
-
-
-class PullingMetricReader(MetricReader):
-    def __init__(
-        self,
-        exporter: PullMetricExporter,
-    ):
-        super().__init__(preferred_temporality=exporter.preferred_temporality)
-        self._exporter = exporter
-        self._exporter._set_collect_callback(self.collect)
-
-    def _receive_metrics(self, metrics: Iterable[Metric]) -> None:
-        if metrics is None:
-            return
-        token = attach(set_value(_SUPPRESS_INSTRUMENTATION_KEY, True))
-        try:
-            self._exporter.export(metrics)
-        except Exception as e:  # pylint: disable=broad-except,invalid-name
-            _logger.exception("Exception while exporting metrics %s", str(e))
-        detach(token)
-
-    def shutdown(self) -> bool:
-        self._exporter.shutdown()
-        return True
